server_game/feat: update attack damage calculator

+++ src/server_game/game/formula/stat_spec.go
@@ -1,5 +1,5 @@
 package formula
 
-func StatSpec_GetPercent(class, id, strengh, dexterity, vitality, energy, leadership int, percent *float32) {
+func StatSpec_GetPercent(class, id, strengh, dexterity, vitality, energy, leadership int, percent *float64) {
 	call(f.StatSpec, "StatSpec_GetPercent", "iiiiiii>d", class, id, strengh, dexterity, vitality, energy, leadership, percent)
 }
